pyproject.toml
[project]
name = "pppp"
version = "0.1.0"
description = "Penultimate Picture Post-Processor"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "setuptools>=65.0.0",
  "httpx>=0.27.0",
  "magika>=1.0.0",
  "paddleocr>=2.7.0,<3.0.0",
  "paddlepaddle>=2.5.0,<3.0.0",
  "pillow>=10.0.0",
    "ram @ git+https://github.com/xinyu1205/recognize-anything.git",
  "torch",
  "torchvision",
  "fairscale",
  "timm==0.4.12",
  "transformers>=4.36,<5",
  "scipy",
]

[dependency-groups]
test = [
    "pytest",
]
dev = [
    "ruff",
    "hatchling"
]

[tool.uv]
package = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
asyncio_mode = "auto"
testpaths = [
    "tests"
]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "target",
    "tests",
]

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # pycodestyle errors
    "E",
    # pycodestyle warnings
    "W",
    # mccabe complexity
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-annotations
    "ANN",
    # flake8-async
    "ASYNC",
    # flake8-bandit
    "S",
    # flake8-blind-except
    "BLE",
    # flake8-boolean-trap
    "FBT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-datetimez
    "DTZ",
    # flake8-debugger
    "T10",
    # flake8-django
    "DJ",
    # flake8-errmsg
    "EM",
    # flake8-executable
    "EXE",
    # flake8-future-annotations
    "FA",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-import-conventions
    "ICN",
    # flake8-logging-format
    "G",
    # flake8-no-pep420
    "INP",
    # flake8-pie
    "PIE",
    # flake8-print
    "T20",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-raise
    "RSE",
    # flake8-return
    "RET",
    # flake8-self
    "SLF",
    # flake8-simplify
    "SIM",
    # flake8-tidy-imports
    "TID",
    # flake8-type-checking
    "TCH",
    # flake8-unused-arguments
    "ARG",
    # flake8-use-pathlib
    "PTH",
    # pandas-vet
    "PD",
    # pygrep-hooks
    "PGH",
    # Pylint
    "PL",
    # tryceratops
    "TRY",
    # flynt
    "FLY",
    # NumPy-specific rules
    "NPY",
    # Airflow
    "AIR",
    # Perflint
    "PERF",
    # refurb
    "FURB",
    # flake8-logging
    "LOG",
    # Ruff-specific rules
    "RUF",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    "PLW2901",
    # Allow boolean positional values in function calls
    "FBT003",
    # Ignore complexity
    "C901",
    # Allow print statements (useful for debugging)
    "T201",
    # Allow assertions in non-test code
    "S101",
    # Allow hardcoded bind all interfaces (0.0.0.0)
    "S104",
    # Allow subprocess without shell=False
    "S603",
    # Allow subprocess with shell=True
    "S602",
    # Allow use of subprocess.run
    "S607",
    # Allow use of assert
    "S101",
    # Allow magic values in comparisons
    "PLR2004",
    # Allow too many arguments
    "PLR0913",
    # Allow too many branches
    "PLR0912",
    # Allow too many statements
    "PLR0915",
    # Allow catching broad exceptions
    "BLE001",
    # Allow long try blocks
    "TRY300",
    # Docstring requirements - we'll handle these selectively
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    # No blank line after docstring (project style prefers a blank line)
    "D202",
    "COM812",
    # Allow f-strings in logging calls
    "G004",
    # Allow camelCase names for variables
    "N815",
    # allow message exceptions
    "TRY003",
    # allow string literals
    "EM101",
    # allow globals
    "PLW0603",
    "PLW0602",
    # literals in exceptions
    "EM102"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools,scripts}/*" = ["E402", "D", "S", "ANN"]
"**/scripts/*" = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-type-checking]
strict = true